fix: return guardoni publicKey as fallback for experiment output
