+++ platforms/guardoni/__mocks__/puppeteer.mock.ts
@@ -7,6 +7,12 @@ const puppeteerMock = puppeteer as jest.Mocked<PuppeteerExtra>;
 const pageMock = {
   on: jest.fn(),
   goto: jest.fn(),
+  waitForSelector: jest.fn().mockImplementation(() => {
+    throw new Error(`waitForSelector: Not implemented`);
+  }),
+  $: jest.fn(),
+  evaluate: jest.fn(),
+  evaluateHandle: jest.fn(),
   waitForTimeout: jest.fn().mockImplementation((ms) => {
     return new Promise((resolve) => {
       setTimeout(resolve, ms / 10);
@@ -16,6 +22,9 @@ const pageMock = {
 const browserMock = {
   pages: jest.fn().mockResolvedValue([pageMock] as any),
 };
-puppeteerMock.launch.mockResolvedValue(browserMock as any);
+puppeteerMock.use.mockImplementation((fn) => puppeteerMock);
+puppeteerMock.launch.mockImplementation(() => {
+  return Promise.resolve(browserMock) as any;
+});
 
 export { puppeteerMock, pageMock, browserMock };
