+++ platforms/guardoni/jest.setup.js
@@ -1,3 +1,7 @@
-process.env.DEBUG = 'guardoni*,-guardoni:debug,@trex*';
-process.env.YT_BACKEND = 'http://localhost:9000/api';
-process.env.TK_BACKEND = 'http://localhost:14000/api';
+import path from 'path';
+
+const envPath = path.resolve(__dirname, '.env.development')
+
+require('dotenv').config({ path: envPath });
+
+process.env.DEBUG = 'guardoni*,@trex*'//-guardoni:debug;
